feat: redesign of menu finalized
