+++ src/lib/components/molecules/MenuLinks.svelte
@@ -4,18 +4,26 @@ import LinkedInIcon from "$lib/icons/LinkedInIcon.svelte";
     let {onCloseMenu} = $props();
 </script>
 
-<a class="icon-link" href="/resume" role="menuitem" onclick={onCloseMenu}>Resume</a>
-<a class="icon-link" href="/goodies" role="menuitem" onclick={onCloseMenu}>Goodies</a>
-<a class="icon-link" target="_blank" rel="noopener noreferrer" href="https://github.com/vojtechmarek-dev" role="menuitem" onclick={onCloseMenu}><GitHubIcon></GitHubIcon>GitHub</a>
-<a class="icon-link" target="_blank" rel="noopener noreferrer" href="https://www.linkedin.com/in/vojtechmarek-dev/" role="menuitem" onclick={onCloseMenu}><LinkedInIcon/>LinkedIn</a>
+<a class="menu-link" href="/resume" role="menuitem" onclick={onCloseMenu}>Resume</a>
+<a class="menu-link" href="/goodies" role="menuitem" onclick={onCloseMenu}>Goodies</a>
+<a class="menu-link" target="_blank" rel="noopener noreferrer" href="https://github.com/vojtechmarek-dev" role="menuitem" onclick={onCloseMenu}><GitHubIcon></GitHubIcon>GitHub</a>
+<a class="menu-link" target="" rel="noopener noreferrer" href="https://www.linkedin.com/in/vojtechmarek-dev/" role="menuitem" onclick={onCloseMenu}><LinkedInIcon/>LinkedIn</a>
 
 <style lang="scss">
-    .icon-link {
+    .menu-link {
             display: flex;
             justify-content: left;
             color: var(--color--text);
             text-decoration: none;
-            gap: 3px;
+            gap: 0.25rem;
             align-items: center;
+            padding: 8px 10px;
+            border-radius: 8px;
+            transition: background-color 0.15s ease, color 0.15s ease;
+            width: 100%;
         }
+
+    .menu-link:hover {
+        background: color-mix(in srgb, var(--color--primary-tint) 50%, transparent);
+    }
 </style>
