+++ src/routes/+layout.svelte
@@ -41,20 +41,14 @@
         color: var(--color--text);
     }
 
-    .page-wrapper {
-        display: flex;
-        flex-direction: column;
-        min-height: 100vh; // Ensure wrapper takes full viewport height
-    }
-
     .main-area {
         flex: 1; // Makes sure this grows to fill available space
-        padding: 30px 100px 100px;
+        padding: 30px 200px 100px;
         position: relative;
         z-index: 3;
 
         @include breakpoints.for-phone-only {
-            padding: 30px 20px 20px;
+            padding: 40px 40px 20px;
         }
     }
 </style>
